refactor: move useSaveUnits into educator api helpers
